fix(react-email): Preview server crashing without React 19
