+++ packages/react-email/src/actions/get-emails-directory-metadata-action.ts
@@ -1,0 +1,19 @@
+'use server';
+
+import type { EmailsDirectory } from '../utils/get-emails-directory-metadata';
+import { getEmailsDirectoryMetadata } from '../utils/get-emails-directory-metadata';
+
+export const getEmailsDirectoryMetadataAction = async (
+  absolutePathToEmailsDirectory: string,
+  keepFileExtensions = false,
+  isSubDirectory = false,
+
+  baseDirectoryPath = absolutePathToEmailsDirectory,
+): Promise<EmailsDirectory | undefined> => {
+  return getEmailsDirectoryMetadata(
+    absolutePathToEmailsDirectory,
+    keepFileExtensions,
+    isSubDirectory,
+    baseDirectoryPath,
+  );
+};
