+++ packages/react-email/src/app/preview/[...slug]/page.tsx
@@ -2,10 +2,10 @@ import path from 'node:path';
 import { Suspense } from 'react';
 import { redirect } from 'next/navigation';
 import { getEmailPathFromSlug } from '../../../actions/get-email-path-from-slug';
-import { getEmailsDirectoryMetadata } from '../../../actions/get-emails-directory-metadata';
 import { renderEmailByPath } from '../../../actions/render-email-by-path';
 import { emailsDirectoryAbsolutePath } from '../../../utils/emails-directory-absolute-path';
 import Home from '../../page';
+import { getEmailsDirectoryMetadata } from '../../../utils/get-emails-directory-metadata';
 import Preview from './preview';
 
 export const dynamicParams = true;
