+++ packages/react-email/src/cli/commands/export.ts
@@ -12,7 +12,7 @@ import { tree } from '../utils';
 import {
   EmailsDirectory,
   getEmailsDirectoryMetadata,
-} from '../../actions/get-emails-directory-metadata';
+} from '../../utils/get-emails-directory-metadata';
 import { renderingUtilitiesExporter } from '../../utils/esbuild/renderring-utilities-exporter';
 
 const getEmailTemplatesFromDirectory = (emailDirectory: EmailsDirectory) => {
