+++ packages/react-email/src/utils/get-emails-directory-metadata.spec.ts
@@ -4,7 +4,7 @@ import { getEmailsDirectoryMetadata } from './get-emails-directory-metadata';
 test('getEmailsDirectoryMetadata on demo emails', async () => {
   const emailsDirectoryPath = path.resolve(
     __dirname,
-    '../../../../apps/demo/emails/',
+    '../../../../apps/demo/emails',
   );
   expect(await getEmailsDirectoryMetadata(emailsDirectoryPath)).toEqual({
     absolutePath: emailsDirectoryPath,
